test: split graphite tests into a module with a shared harness
Converts the flat tests/integration/graphite.rs into a graphite/ module so
the gt runner can be shared with the upcoming remote-backed test suite.

  - graphite/graphite_test_harness.rs: the gt() runner, git-shim helpers,
    require_gt!, and the shared setup/assert helpers, lifted verbatim.
  - graphite/local_ops.rs: the existing tests, unchanged.

main.rs needs no edit: `mod graphite;` resolves to graphite/mod.rs. Test
paths shift to graphite::local_ops::*, so TEST_FILTER=graphite still
selects the whole suite.

No behavior change; all 21 non-ignored tests pass as before.
